app/requirements.txt
openai
pydub
aiogram
tiktoken
sqlalchemy
python-dotenv
